Consolidate statement tests
