--- test/syntaxes/lock-statements.test.syntax.ts
@@ -1,99 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-
-import { should } from 'chai';
-import { tokenize, Input, Token } from './utils/tokenize';
-
-describe("Grammar", () => {
-    before(() => should());
-
-    describe("Lock statements", () => {
-        it("single-line lock with embedded statement", () => {
-            const input = Input.InMethod(`lock (new object()) Do();`);
-            const tokens = tokenize(input);
-
-            tokens.should.deep.equal([
-                Token.Keywords.Lock,
-                Token.Punctuation.OpenParen,
-                Token.Keywords.New,
-                Token.PrimitiveType.Object,
-                Token.Punctuation.OpenParen,
-                Token.Punctuation.CloseParen,
-                Token.Punctuation.CloseParen,
-                Token.Identifiers.MethodName("Do"),
-                Token.Punctuation.OpenParen,
-                Token.Punctuation.CloseParen,
-                Token.Punctuation.Semicolon
-            ]);
-        });
-
-        it("single-line lock with block", () => {
-            const input = Input.InMethod(`lock (new object()) { Do(); }`);
-            const tokens = tokenize(input);
-
-            tokens.should.deep.equal([
-                Token.Keywords.Lock,
-                Token.Punctuation.OpenParen,
-                Token.Keywords.New,
-                Token.PrimitiveType.Object,
-                Token.Punctuation.OpenParen,
-                Token.Punctuation.CloseParen,
-                Token.Punctuation.CloseParen,
-                Token.Punctuation.OpenBrace,
-                Token.Identifiers.MethodName("Do"),
-                Token.Punctuation.OpenParen,
-                Token.Punctuation.CloseParen,
-                Token.Punctuation.Semicolon,
-                Token.Punctuation.CloseBrace
-            ]);
-        });
-
-        it("lock with embedded statement", () => {
-            const input = Input.InMethod(`
-lock (new object())
-    Do();`);
-            const tokens = tokenize(input);
-
-            tokens.should.deep.equal([
-                Token.Keywords.Lock,
-                Token.Punctuation.OpenParen,
-                Token.Keywords.New,
-                Token.PrimitiveType.Object,
-                Token.Punctuation.OpenParen,
-                Token.Punctuation.CloseParen,
-                Token.Punctuation.CloseParen,
-                Token.Identifiers.MethodName("Do"),
-                Token.Punctuation.OpenParen,
-                Token.Punctuation.CloseParen,
-                Token.Punctuation.Semicolon
-            ]);
-        });
-
-        it("lock with block", () => {
-            const input = Input.InMethod(`
-lock (new object())
-{
-    Do();
-}`);
-            const tokens = tokenize(input);
-
-            tokens.should.deep.equal([
-                Token.Keywords.Lock,
-                Token.Punctuation.OpenParen,
-                Token.Keywords.New,
-                Token.PrimitiveType.Object,
-                Token.Punctuation.OpenParen,
-                Token.Punctuation.CloseParen,
-                Token.Punctuation.CloseParen,
-                Token.Punctuation.OpenBrace,
-                Token.Identifiers.MethodName("Do"),
-                Token.Punctuation.OpenParen,
-                Token.Punctuation.CloseParen,
-                Token.Punctuation.Semicolon,
-                Token.Punctuation.CloseBrace
-            ]);
-        });
-    });
-});
