--- test/syntaxes/yield-statements.test.syntax.ts
@@ -1,36 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-
-import { should } from 'chai';
-import { tokenize, Input, Token } from './utils/tokenize';
-
-describe("Grammar", () => {
-    before(() => should());
-
-    describe("Yield statements", () => {
-        it("yield return", () => {
-            const input = Input.InMethod(`yield return 42;`);
-            const tokens = tokenize(input);
-
-            tokens.should.deep.equal([
-                Token.Keywords.Control.Yield,
-                Token.Keywords.Control.Return,
-                Token.Literals.Numeric.Decimal("42"),
-                Token.Punctuation.Semicolon
-            ]);
-        });
-
-        it("yield break", () => {
-            const input = Input.InMethod(`yield break;`);
-            const tokens = tokenize(input);
-
-            tokens.should.deep.equal([
-                Token.Keywords.Control.Yield,
-                Token.Keywords.Control.Break,
-                Token.Punctuation.Semicolon
-            ]);
-        });
-    });
-});
